# src/main/java/dev/hanluc/expensetracker/expense/domain/vo/Money.java
package dev.hanluc.expensetracker.expense.domain.vo;

import jakarta.persistence.Embeddable;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotNull;

@Embeddable
public record Money(

  @NotNull
  @Min(1)
  Long value,

  @NotNull
  Integer exponent

) { }